src/models/requirements.txt
azure-ai-projects>=2.0.0
azure-identity
python-dotenv
tiktoken
